Update game visuals and adjustment to game physics
The commit includes changes to both CSS and image assets to improve the visual aspect of the game. Most images have been moved to the assets folder for better organization. Adjustments to the game-container have been made as well. In addition to visual updates, some changes have been made to the game physics specifically to the acceleration factor and acceleration max in DoodleJump.ts. The Canvas Doodle Renderer has been updated to correctly render the game elements based on the canvas ratio.