Adjust padding in PlayerModal and change date formatting
This commit reduces the padding-bottom of 'player-modal__score-title' and 'player-modal__name-title' in styles.css for a tighter layout. It also changes how dates are saved in Firebase and displayed in App from ISO string to locale string for greater readability. A default player is used if no player data is already present to improve user experience.